Move to use interface model to allow either model in use from the same version of the package. Fix tests because of static version/resource access
